chore: remove checking count
The API documents the count as only being approximate and its
value is not accurate enough to use in the tests.  When the
tests were running I watched the error reporting console and saw
the count be (inccorectly) 2 for a few seconds before dropping
down to the correct value of 1.