chore: increase system test timeout
It looks like the underlying issue with the system tests is that
it may take longer than the test timeout before the API reports
an error.  I have increased the timeout to 10 minutes since the
code will now poll the server so that it can quit early if a
response comes back earlier than the timeout.